Move testing errors into its own test class
